@paradxum requested changes -- signals.py
