+++ simple_sso/signals.py
@@ -1,8 +1,18 @@
 
-from simple_sso.settings import settings
 from django.urls import NoReverseMatch, reverse
-
+from simple_sso.settings import settings
+from simple_sso.sso_server.models import Token
 from webservices.sync import SyncConsumer
+
+
+def logout_token(sender, request, **kwargs):
+    """
+    A signal receiver which removes a token when its users logs out.
+    """
+    tokens = Token.objects.select_related('session').filter(session__session_key=request.session.session_key)
+
+    for token in tokens:
+        token.delete()
 
 
 def logout_sso_client(sender, request, user, **kwargs):
