+++ simple_sso/sso_client/apps.py
@@ -6,6 +6,6 @@ class SimpleSSOClient(AppConfig):
     name = 'simple_sso.sso_client'
 
     def ready(self):
-        from .models import logout_sso_client
+        from simple_sso.signals import logout_sso_client
         user_logged_out.connect(
             logout_sso_client, dispatch_uid='logout_sso_client')
