+++ simple_sso/sso_server/apps.py
@@ -6,5 +6,5 @@ class SimpleSSOServer(AppConfig):
     name = 'simple_sso.sso_server'
 
     def ready(self):
-        from .models import logout_token
+        from simple_sso.signals import logout_token
         user_logged_out.connect(logout_token, dispatch_uid='logout_token')
